add first person playershadow ips + cheatengine table
This patch alters the thirdperson game logic in a way that it spawns the player model but keeps all the first person logic that would be disabled otherwise. Visual clipping with the thirdperson player model and the first person camera can be fixed by adding the "player model" flag to the player model and all other viewmodels such as guns and hands.

squashed two commits to ignore .gitignore (everything) :o